requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
pyserial==3.5
numpy==1.24.3
scipy==1.11.1
python-socketio==5.8.0
eventlet==0.33.3
